Change or remove punctuation mark from sentence
In using this library, I thought it would be nice to add in an option to change and/or remove the punctuation mark at the end of a sentence. This would make it ideal for generating random things like alert messages that you may want to have end with an exclamation mark instead of a period, or page titles where you may want no punctuation mark at all.